renaming the files so that removing the new from the files and the folder
