+++ Examples/src/components/AppDeatilsRouters/AppDeatilsRouter.scss
